requirements.txt
fastapi
uvicorn
tensorflow-lite
numpy